Store type of index in the catalog. Exercise 7.14
Add test as well

Easy to miss important additions. Needed to extend many places in the code to store this. For example,
there are a few constructors of the IndexInfo, and if a constructor was
missing the intialization of the indextype field, it would be a silent
error; just printing out "" instead of the index type, because nothing
was actually stored.