fix uncaught error when trying to work with a table that does not exist
Previously, I had a typo in the indexupdatetest where I was looking for
a table that did not exist in the database. The only error that would
bubble up was that the offset in the RecordPage was negative and that
is a fatal error. Eventually, I found the error was because I was
looking for a table that did not exist.

Now, I catch this error in the tablemanager. If the table is not found
during the scan of the table catalog, an exception is raised there with
a clearer error message